backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=1.0.0
playwright>=1.49.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
google-cloud-logging>=3.11.0
httpx>=0.27.0
google-adk>=1.0.0
reportlab>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
